Add pow_assent api user flow
Using PowAssent to create handle user authentication, login, and
registration.  If we want to add PowAssent Oauth with Github, we will
need to do some tweaking as it expects an html page to be rendered
rather than an api call.  I also had to move pow config to config.exs
instead of runtime.exs... but this may be able to change with some
tweaking.